review comments p1
