+++ src/components/nimbleTimeline/StyleWrappers.tsx
@@ -1,0 +1,72 @@
+import {styled} from '@mui/system';
+
+interface GroupHeaderProps {
+  headerWidth: string | number | undefined;
+}
+
+const ItemContent = styled('div')({
+  display: 'flex',
+  justifyContent: 'center',
+  width: 'inherit',
+  border: 'none',
+  overflow: 'hidden',
+  whiteSpace: 'nowrap',
+  textOverflow: 'ellipsis',
+});
+
+const GroupContainer = styled('div')({
+  display: 'flex',
+  height: 'inherit',
+});
+
+const GroupLeftSection = styled('div')({
+  display: 'flex',
+  width: '50%',
+  borderRight: '1px solid #9A9FA5',
+  fontSize: '14px',
+  alignItems: 'center',
+  overflow: 'scroll',
+});
+
+const GroupRightSection = styled('div')({
+  display: 'flex',
+  height: 'inherit',
+  padding: '4px 0px',
+  width: '50%',
+  overflow: 'scroll',
+  alignItems: 'center',
+});
+
+const GroupHeaderContainer = styled('div')<GroupHeaderProps>(props => ({
+  display: 'flex',
+  width: props.headerWidth,
+  background: 'white',
+  color: 'black',
+}));
+
+const GroupHeaderLeft = styled('div')({
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+  width: '50%',
+  fontSize: '12px',
+  borderRight: '1px solid #9A9FA5',
+});
+
+const GroupHeaderRight = styled('div')({
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+  width: '50%',
+  fontSize: '12px',
+});
+
+export {
+  ItemContent,
+  GroupContainer,
+  GroupLeftSection,
+  GroupRightSection,
+  GroupHeaderContainer,
+  GroupHeaderLeft,
+  GroupHeaderRight,
+};
