+++ src/components/nimbleTimeline/timelineStyles.css
@@ -1,5 +1,10 @@
 @import url('https://fonts.googleapis.com/css?family=Work+Sans');
 
+*::-webkit-scrollbar {
+    width: 0.01em;
+    height: 0.2em;
+}
+
 body {
     font-family: 'Work Sans', sans-serif;
     font-size: 18px;
